group state records
